Update read_meta.py
After looking at the sample commands that Tyson posted on the Wiki, I added a function that will convert the .tif files stored in the specified directory from whatever coordinate system used to the projection in use by DAYMET. I haven't been able to test any of the output files yet though. The script appends .converted.tif to the end of the original files.